loading of joystick config at joystick detection
